Add infrastructure for aborting (#2497)
I've messed around with cancellation patterns a bit but figured since the JS world is moving toward AbortController we should give that a go.

It's not the nicest API but does work.  This commit adds some utilities that make use more convenient and deletes some of my competing experiments that we never actually used.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>