New "observable value" primitive (#2642)
This is a new class that does a better job of marrying `Observable` and `Promise` semantics.  It is similar to gate but with the `Observable` interface so you can listen for an event repeatedly.  This would theoretically make a good replacement for events like `NodeLifecycle#online` so you can await them without checking e.g. `NodeLifecycle.#isOnline` but not making that move for now.  I did use it to implement new `ObservableSet#empty` event.